Pluggable Annotation class
We're already using the Flask.g to store session variables, hooks, and the Annotation class itselft can be a session variable too.

That way the class itself is pluggable so custom implementation can be used within our framework.
If no custom implementation is set, we set the default Annotation class.